+++ js/main.js
@@ -1,9 +1,11 @@
-import {startValidation} from './form.js';
-import {createMap} from './map.js';
 import {getData} from './server.js';
+import {createMap} from './map.js';
+import {onResetEvent} from './form.js';
+import {startValidation} from './validation.js';
 
 getData((offers) => {
   createMap(offers);
 });
 
+onResetEvent();
 startValidation();
