--- js/popup.js
@@ -1,2 +1,0 @@
-// Модуль работы с подробной информацией об объявлении.
-// Пункт ТЗ: 5.4, 5.5, 5.6, 5.7.
